--- autoforce/core/modules/regressor_md/__init__.py
@@ -1,5 +1,0 @@
-# +
-from .kernel_regressor import KernelRegressor_md
-from .regressor import Regressor_md
-
-__all__ = ["Regressor_md", "KernelRegressor_md"]
