+++ autoforce/functions/bijections/range_bi.py
@@ -7,14 +7,6 @@ from torch import Tensor
 from autoforce.core.functions import Bijection_fn
 
 
-class Identity_bi(Bijection_fn):
-    def function(self, x: Tensor) -> Tensor:
-        return x
-
-    def inverse(self, x: Tensor) -> Tensor:
-        return x
-
-
 class Range_bi(Bijection_fn):
     def __init__(self, a: float, b: float) -> None:
         assert b > a
