+++ autoforce/functions/kernels/__init__.py
@@ -1,0 +1,4 @@
+# +
+from .dotprod_kern import DotProd_kern
+
+__all__ = ["DotProd_kern"]
