+++ autoforce/functions/kernels/dotprod_kern.py
@@ -6,6 +6,6 @@ from torch import Tensor
 from autoforce.core.functions import Kernel_fn
 
 
-class DotProductKernel(Kernel_fn):
+class DotProd_kern(Kernel_fn):
     def function(self, uv: Tensor, u: Tensor, v: Tensor) -> Tensor:
         return uv / (u * v)
